move function to common_functionality
