+++ editor/src/messages/tool/common_functionality/utility_functions.rs
@@ -4,7 +4,7 @@ use crate::messages::tool::common_functionality::graph_modification_utils::get_t
 use glam::DVec2;
 use graphene_core::renderer::Quad;
 use graphene_core::text::{FontCache, load_face};
-use graphene_std::vector::PointId;
+use graphene_std::vector::{ManipulatorPointId, PointId, SegmentId, VectorData};
 
 /// Determines if a path should be extended. Goal in viewport space. Returns the path and if it is extending from the start, if applicable.
 pub fn should_extend(
@@ -65,4 +65,31 @@ pub fn text_bounding_box(layer: LayerNodeIdentifier, document: &DocumentMessageH
 	let far = graphene_core::text::bounding_box(text, buzz_face.as_ref(), typesetting, false);
 
 	Quad::from_box([DVec2::ZERO, far])
+}
+
+pub fn calculate_segment_angle(anchor: PointId, segment: SegmentId, vector_data: &VectorData, pen_tool: bool) -> Option<f64> {
+	let is_start = |point: PointId, segment: SegmentId| vector_data.segment_start_from_id(segment) == Some(point);
+	let anchor_position = vector_data.point_domain.position_from_id(anchor)?;
+	let end_handle = ManipulatorPointId::EndHandle(segment).get_position(vector_data);
+	let start_handle = ManipulatorPointId::PrimaryHandle(segment).get_position(vector_data);
+
+	let start_point = if is_start(anchor, segment) {
+		vector_data.segment_end_from_id(segment).and_then(|id| vector_data.point_domain.position_from_id(id))
+	} else {
+		vector_data.segment_start_from_id(segment).and_then(|id| vector_data.point_domain.position_from_id(id))
+	};
+
+	let required_handle = if is_start(anchor, segment) {
+		start_handle
+			.filter(|&handle| pen_tool && handle != anchor_position)
+			.or(end_handle.filter(|&handle| Some(handle) != start_point))
+			.or(start_point)
+	} else {
+		end_handle
+			.filter(|&handle| pen_tool && handle != anchor_position)
+			.or(start_handle.filter(|&handle| Some(handle) != start_point))
+			.or(start_point)
+	};
+
+	required_handle.map(|handle| -(handle - anchor_position).angle_to(DVec2::X))
 }
